+++ src/graphql/export.rs
@@ -708,8 +708,8 @@ where
 {
     let (start, end) = filter.time();
     let iter = store.boundary_iter(
-        &lower_closed_bound_key(key_prefix, start),
-        &upper_open_bound_key(key_prefix, end),
+        &lower_closed_bound_key(Some(key_prefix), start),
+        &upper_open_bound_key(Some(key_prefix), end),
         Direction::Forward,
     );
     export_file(iter, filter, export_type, export_path)